requirements.txt
Flask
Flask-Mail
bcrypt
python-dotenv
pyodbc